ref: isolate working with functions to avoid __future__ conflicts
